+++ NEWS.md
@@ -3,6 +3,10 @@
 This release contains a fix to the issue #64: update to code to be compatible
 with ggplot2 v4.0.0 and previous version.
 
+- add `log_scale` parameter in the `plot_step_ahead_model_output()` function to
+allow log scale on the y-axis
+- fix shared axis on interactive plot version
+
 # hubVis 0.1.2
 
 This release contains a fix to the issue #56: fix palette creation for `*_color`